tinggal finishing

+++ home/homeadmin.php
@@ -5,8 +5,14 @@
     if(isset($_SESSION['id']) == ""){
         header("Location: index.php");
     }else{
-        $info = mysqli_query($conn, "SELECT * FROM mahasiswa");
+        $sql = mysqli_query($conn, "SELECT * FROM mahasiswa");
     }
+
+    if(isset($_GET['search'])){
+        $nama = $_GET['search'];
+        $sql = mysqli_query($conn, "SELECT * FROM mahasiswa WHERE nama LIKE '%$nama%'");                        
+    }
+    
 ?>
 
 <html>
@@ -18,6 +24,14 @@
     <link href="https://cdn.jsdelivr.net/npm/bootstrap@5.3.0-alpha1/dist/css/bootstrap.min.css" rel="stylesheet" integrity="sha384-GLhlTQ8iRABdZLl6O3oVMWSktQOp6b7In1Zl3/Jr59b6EGGoI1aFkw7cmDA6j6gD" crossorigin="anonymous">
     <script src="https://cdn.jsdelivr.net/npm/bootstrap@5.3.0-alpha1/dist/js/bootstrap.bundle.min.js" integrity="sha384-w76AqPfDkMBDXo30jS1Sgez6pr3x5MlQ1ZAGC+nuZB+EYdgRZgiwxhTBTkF7CXvN" crossorigin="anonymous"></script>
     <title>Portal Akademik UAS</title>
+
+    <style>
+        .contrainer{
+            display: flex;
+            justify-content: center;
+            align-items: center;
+        }
+    </style>
 </head>
 
 <body>
@@ -30,22 +44,16 @@
         <div class="collapse navbar-collapse" id="navbarSupportedContent">
         <ul class="navbar-nav me-auto mb-2 mb-lg-0">
         </ul>
-        <form class="d-flex" role="search">
-            <input class="form-control me-2" type="search" placeholder="Pencarian Mahasiswa" aria-label="Search">
+        <form class="d-flex" role="search" action="homeadmin.php" method="GET">
+            <input class="form-control me-2" type="search" name="search" placeholder="Pencarian berdasarkan Nama" aria-label="Search">
             <button class="btn btn-outline-dark me-2" type="submit">Search</button>
             <a href="../action/logout.php" class="btn btn-primary">Logout</a>
         </form>
         </div>
     </div>
     </nav>
-        <!-- <a href="tambah.php">Tambah Data Baru</a> | <a href="cetak_mhs.php">Cetak Mahasiswa</a><br /><br /><br /><br /> -->
-        <!-- <a href="index.php">Logout</a> -->
-        <!-- <form action="cari.php" method="get">
-            <input type="text" name="cari" id="cari" placeholder="Pencarian berdasarkan nama...">
-        </form> -->
-
-
-        <table class="table table-bordered">
+    <div class="container">
+    <table class="table table-bordered">
             <tr>
                 <th>NIM</th>
                 <th>Nama</th>
@@ -55,20 +63,25 @@
                 <th>Action</th>
             </tr>
                 <?php
-                    while($user_krs = mysqli_fetch_array($info)){
-                        echo "<tr><td>" . $user_krs['nim'] . "</td>";
-                        echo "<td>" . $user_krs['nama'] . "</td>";
-                        echo "<td>" . $user_krs['jenis_kelamin'] . "</td>";
-                        echo "<td>" . $user_krs['alamat'] . "</td>";
-                        echo "<td>" . $user_krs['tgl_lahir'] . "</td>";
-                        echo "<td><a class='btn btn-primary' href='../action/edit.php?id=$user_krs[id]'>Edit </a> | <a class='btn btn-primary' href='../action/delete.php?id=$user_krs[id]'>Delete</a></td></tr>";
+                    if(mysqli_num_rows($sql) > 0){
+                        while($user_krs = mysqli_fetch_array($sql)){
+                            echo "<tr><td>" . $user_krs['nim'] . "</td>";
+                            echo "<td>" . $user_krs['nama'] . "</td>";
+                            echo "<td>" . $user_krs['jenis_kelamin'] . "</td>";
+                            echo "<td>" . $user_krs['alamat'] . "</td>";
+                            echo "<td>" . $user_krs['tgl_lahir'] . "</td>";
+                            echo "<td><a class='btn btn-primary' href='../action/edit.php?id=$user_krs[id]'>Edit </a> | <a class='btn btn-primary' href='../action/delete.php?id=$user_krs[id]'>Delete</a></td></tr>";
+                        }
+                    }else{
+                        echo "<tr><td colspan='6' align='center'> Data Tidak Ditemukan!!! </td></tr>";
                     }
+
                 ?>
             <tr>
                 <td colspan="6"><a class="btn btn-primary" href="../action/tambahuser.php">Tambah Akun</a></td>
             </tr>
         </table>
- 
+    </div>
 </body>
 
 </html>
